Pt_6_1: Top 10 most popular songs of your favorite artist according to Yandex Music
A parser program for extracting the top 10 most popular songs of your favorite artist from a web resource according to Yandex Music. The user must submit a link to the selected artist on this web resource for input. In the case of an incorrectly specified link or other incorrect operation scenario, error handling is provided.